moved css file to the right place. Adjustments for survey to mirror text scenarios
