+++ dashboard-ui/src/components/Survey/comparison.jsx
@@ -3,7 +3,7 @@ import { ElementFactory, Question, Serializer } from "survey-core";
 import { SurveyQuestionElementBase } from "survey-react-ui";
 import { Button, Modal } from "react-bootstrap";
 import Dynamic from "./dynamic";
-import './template.css'
+import '../../css/template.css';
 import { useSelector } from "react-redux";
 
 const CUSTOM_TYPE = "comparison";
