+++ dashboard-ui/src/css/template.css
@@ -10,9 +10,8 @@
 
 .scene {
   display: flex;
-  flex-direction: row;
+  flex-direction: column;
   width: 100%;
-  justify-content: space-between;
 }
 
 .patient-buttons {
